[Win] Fix uncommon not expanding on search
Fixes #504
