+++ Xamarin.PropertyEditing.Windows/Xamarin.PropertyEditing.Windows.csproj
@@ -99,6 +99,7 @@
     <Compile Include="CurrentColorEditorControl.cs" />
     <Compile Include="DoubleToPercentageConverter.cs" />
     <Compile Include="DoubleToQuantityConverter.cs" />
+    <Compile Include="FilterExpander.cs" />
     <Compile Include="HasItemsToVisibilityConverter.cs" />
     <Compile Include="InvertedVisibilityConverter.cs" />
     <Compile Include="IPropertiesHost.cs" />
